wrapping up for the year

+++ Codeforces/pinely Round 3/B_Make_Almost_Equal_With_Mod.cpp
@@ -15,7 +15,7 @@ void solve()
     ll fact = 2;
     while(mt2)
     {
-        unordered_map<int,int> m;
+        unordered_map<ll,ll> m;
         for(auto x:a)
         {
             m[x%fact]++;
